requirements.txt
streamlit
pypdf
google.generativeai
streamlit
OpenAI